requirements.txt
feedparser>=6.0.10
anthropic>=0.18.1
openai>=1.13.3
google-api-python-client>=2.118.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
pydub>=0.25.1
PyYAML>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
audioop-lts>=0.2.1
beautifulsoup4>=4.12.0
httpx
PyGithub
pytest
pytest-mock
google-auth-oauthlib>=1.2.0
